Cleanup the logic of handling description links
